+++ frontend/html/layout.html
@@ -16,7 +16,7 @@
         }
     </script>
 
-    {% block rss %}
+    {% block feeds %}
     <link rel="alternate" type="application/rss+xml"
           title="Вастрик.Клуб: Новые посты"
           href="{{ settings.APP_HOST }}{% url "rss" %}" />
