+++ frontend/html/users/profile.html
@@ -10,7 +10,7 @@
     {{ user.full_name }} — {{ block.super }}
 {% endblock %}
 
-{% block rss %}
+{% block feeds %}
     <link rel="alternate" type="application/rss+xml"
           title="Вастрик.Клуб: Посты {{ user.slug }}"
           href="{{ settings.APP_HOST }}{% url "user_rss" user.slug %}" />
